+++ tests/test_commands/test_destroy_vm.py
