+++ tests/test_commands/test_disconnect_vm.py
