+++ tests/test_commands/test_get_vnic_from_vm.py
