+++ tests/test_commands/test_power_management_vm.py
